Split the tasks up for probably no good reason... it may become obvious later.
